feat: adding scripts for notifications plus utils related functions

+++ src/utils/date_utils.py
@@ -66,7 +66,7 @@ def get_formatted_date(date: str) -> datetime:
     return datetime.strptime(date[:-6], "%Y-%m-%dT%H:%M:%S")
 
 
-def is_time_in_surrounding_hours(check_time: time, hours: int) -> bool:
+def is_time_in_surrounding_hours(check_time: time, hours: float) -> bool:
     utc_time_now = datetime.utcnow()
     begin_time = (utc_time_now - timedelta(hours=hours)).time()
     end_time = (utc_time_now + timedelta(hours=hours)).time()
